Log in screen
At start of application log in screen is created, with obvious
functionality. Statusbar is updated on submission.
For those that do the checking/hashing of log in info, you can access
the info using the getUsername() and getPassword() functions. Processing
of the event of submitting can be handled either in the OnSubmit()
function in LogIn.cpp or in main.cpp (which might be best so you can
alter the statusbar again when you know if it was a valid login)